requirements.txt
Django==1.10.1
django-recaptcha==1.1.0
django-vote==2.0.1
